Extra Credit - New Paths
Created additional waves and paths. Path 0 is a parabola, path 1 is a zig-zag across the screen, and path 2 is a zig-zag inwards, down, across, up, and then off the screen.

They can be found in the "Waves and Paths" folder, but will not be visible in-game until we start adding more enemy waves.